pyproject.toml
[project]
name = "canaimeapi"
version = "0.1.0"
description = "API para extrair dados do sistema Canaimé"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "playwright>=1.37.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["canaimeapi"]

[tool.ruff]
line-length = 88
target-version = "py313"
